lambda/requirements.txt
torch==2.1.0
transformers==4.35.0
Pillow==10.1.0
numpy==1.24.3
boto3==1.34.0
requests==2.31.0